se agrega confirmacion antes de eliminar un registro

+++ resources/js/Pages/Area.vue
@@ -88,7 +88,9 @@ const Edit = async (data) => {
 
 const Delete  = async (data) => {
     
-    let resp = await axios.delete(route('delete.area', data.id));    
+    if (confirm('¿Estas seguro de eliminar este registro?')) {
+
+        let resp = await axios.delete(route('delete.area', data.id));    
     
         if (resp.data.result == 1) {
             showSuccess(resp.data.msg)
@@ -96,6 +98,9 @@ const Delete  = async (data) => {
         } else {
             showError(resp.data.msg);            
         }
+    } else {
+        return false;
+    }
 }
 
 const ClearForm = () => {
