+++ resources/js/Pages/Productos.vue
@@ -152,13 +152,16 @@ const Edit = async (data) => {
 }
 
 const Delete  = async (data) => {
-    
-    let resp = await axios.delete(route('delete.producto', data.id));
-    if (resp.data.result == 1) {        
-        showSuccess(resp.data.msg)
-        router.reload({ only: ['Productos'] });
+    if (confirm('¿Estas seguro de eliminar este registro?')) {    
+        let resp = await axios.delete(route('delete.producto', data.id));
+        if (resp.data.result == 1) {        
+            showSuccess(resp.data.msg)
+            router.reload({ only: ['Productos'] });
+        } else {
+            showError(resp.data.msg);            
+        }
     } else {
-        showError(resp.data.msg);            
+        return false;
     }
 }
