+++ resources/js/Pages/Usuarios.vue
@@ -94,12 +94,18 @@ const Edit = async (data) => {
 }
 
 const Delete = async (data) => {
-    let resp = await axios.delete(route('delete.usuario', data.id));
-    if(resp.data.result == 1){
-        showSuccess(resp.data.msg);
-        router.reload({only:['users']});
-    }else{
-        showError(resp.data.msg);
+    if (confirm('¿Estas seguro de eliminar este registro?')) {
+        
+        let resp = await axios.delete(route('delete.usuario', data.id));
+        if(resp.data.result == 1){
+            showSuccess(resp.data.msg);
+            router.reload({only:['users']});
+        }else{
+            showError(resp.data.msg);
+        }
+
+    } else {
+        return false;
     }
 }
